// packages/sms-core/src/core/domain/capabilities/sms-capability.enum.ts
export enum SmsCapability {
  SEND_SINGLE = 'SEND_SINGLE',
  SEND_BULK = 'SEND_BULK',
  SEND_LIKE_TO_LIKE = 'SEND_LIKE_TO_LIKE',
  SEND_PATTERN = 'SEND_PATTERN',
  GET_DELIVERY = 'GET_DELIVERY',
  GET_BALANCE = 'GET_BALANCE',
  GET_LINES = 'GET_LINES',
  RECEIVE_MESSAGES = 'RECEIVE_MESSAGES',
  CANCEL_SCHEDULED = 'CANCEL_SCHEDULED',
  WEBHOOK = 'WEBHOOK',
}